Email Author Module/requirements.txt
nltk==3.2.1
pyenchant==1.6.7
